requirements.txt
streamlit>=1.33.0
google-genai>=0.3.0
python-dotenv>=1.0.0